+++ src/services/application/notification-builder/index.ts
@@ -1,2 +1,1 @@
 export * from './notification.builder';
-export * from './notification.builder.factory.provider';
